@@ -16,6 +16,7 @@
     "@emotion/styled": "^11.11.0",
     "@phosphor-icons/react": "^2.0.15",
     "@testing-library/react": "^15.0.5",
+    "axios": "^1.6.8",
     "formik": "^2.4.6",
     "framer-motion": "^11.0.23",
     "jest": "^29.7.0",
@@ -23,7 +24,8 @@
     "react": "^18.3.1",
     "react-dom": "^18",
     "react-test-renderer": "^18.3.1",
-    "ts-jest": "^29.1.2"
+    "ts-jest": "^29.1.2",
+    "yup": "^1.4.0"
   },
   "devDependencies": {
     "@babel/core": "^7.24.4",
@@ -34,10 +36,10 @@
     "@types/react-dom": "^18",
     "autoprefixer": "^10.0.1",
     "babel-jest": "^29.7.0",
-    "eslint": "^8",
+    "eslint": "^8.57.0",
     "eslint-config-next": "14.1.4",
     "postcss": "^8",
-    "tailwindcss": "^3.3.0",
+    "tailwindcss": "^3.4.3",
     "typescript": "^5"
   }
 }
