@@ -1,116 +1,130 @@
-import React, { useState } from "react";
-import Link from "next/link";
-import { Eye, EyeSlash } from "@phosphor-icons/react";
-import { GoogleLogo, AppleLogo, FacebookLogo } from "@phosphor-icons/react";
-import { API_BASE_URL, ACCESS_TOKEN_NAME } from "../../constants/apiConstants";
-import { error } from "console";
-const axios = require("axios");
+import React, { useState } from 'react';
+import Link from 'next/link';
+import { Eye, EyeSlash } from '@phosphor-icons/react';
+import { useFormik } from 'formik';
+import * as Yup from 'yup';
+import { API_BASE_URL } from '../../constants/apiConstants';
+import axios from 'axios';
+import { useRouter } from "next/router";
+
+const LoginSchema = Yup.object().shape({
+  email: Yup.string().email('Invalid email').required('Email is required'),
+  password: Yup.string().required('Password is required'),
+});
 
 const Login: React.FC = () => {
-  const [email, setEmail] = useState("");
-  const [password, setPassword] = useState("");
+
+  const [errorMessages, setErrorMessages] = useState("");
   const [showPassword, setShowPassword] = useState(false);
-  const [visibleInvalidFields, setVisibleInvalidFields] = useState(false);
-  const [errorMessage, setErrorMessage] = useState("your mother");
   const togglePasswordVisibility = () => setShowPassword(!showPassword);
+  const router = useRouter();
 
-  const handleEmailChange = (e: React.ChangeEvent<HTMLInputElement>) => {
-    setEmail(e.target.value);
-  };
+  const formik = useFormik({
+    initialValues: {
+      email: '',
+      password: '',
+    },
+    validationSchema: LoginSchema,
+    onSubmit: async (values) => {
+      try {
+        const response = await axios.post(`${API_BASE_URL}/api/auth/signin`, {
+          username: values.email,
+          email: values.email,
+          password: values.password,
+        });
 
-  const handlePasswordChange = (e: React.ChangeEvent<HTMLInputElement>) => {
-    setPassword(e.target.value);
-  };
-
-  const handleEmailLogin = async () => {
-    // Handle email login
-    console.log(email, password);
-
-    axios
-      .post(API_BASE_URL + "/api/auth/signin", {
-        username: email,
-        password: password,
-      })
-      .then((response: any) => {
-        console.log(response);
-        if (response.status != 201) {
-        } else {
-          setVisibleInvalidFields(false);
+        if (response.status === 201) {
+          console.log('Login successful', response.data);
+          // Redirect to dashboard or perform other success actions
+          router.push("/");
         }
-      })
-      .catch((error: any) => {
-        console.log(error.response.data.message);
-        setVisibleInvalidFields(true);
-        setErrorMessage(error.response.data.message);
-      });
-  };
+      } catch (error: any) {
+        console.error('Login failed', error.response?.data?.message);
+        setErrorMessages(error.response?.data?.message);
+        console.log("Error Message is " + errorMessages);
+      }
+    },
+  });
 
   return (
-    <div className="flex flex-col items-center justify-center min-h-[90vh] bg-white">
+    <div className="flex flex-col items-center justify-center min-h-[90vh] bg-white font-inter">
       <div className="p-8 bg-gray-100 shadow-sm rounded-xl w-96 h-96">
         <h2 className="text-2xl font-semibold mb-4 text-gray-800">Log In</h2>
-        <div className="mb-4">
-          <label
-            htmlFor="email"
-            className="block text-sm font-medium text-gray-700"
-          >
-            Email
-          </label>
-          <input
-            id="email"
-            type="email"
-            onChange={handleEmailChange}
-            className="mt-1 px-4 py-2 block w-full border-gray-300 rounded-md shadow-sm text-sm text-gray-800"
-          />
-        </div>
-        <div className="mb-4 relative">
-          <label
-            htmlFor="password"
-            className="block text-sm font-medium text-gray-700"
-          >
-            Password
-          </label>
-          <input
-            id="password"
-            type={showPassword ? "text" : "password"}
-            value={password}
-            onChange={handlePasswordChange}
-            className="mt-1 px-4 py-2 block w-full border-gray-300 rounded-md shadow-sm text-sm text-gray-800"
-          />
-          <div className="absolute inset-y-0 right-0 pr-3 flex items-center text-sm leading-5">
-            <button onClick={togglePasswordVisibility}>
-              {showPassword ? <EyeSlash size={20} /> : <Eye size={20} />}
+        <form onSubmit={formik.handleSubmit}>
+          {/* Email Input */}
+          <div className="mb-4">
+            <label htmlFor="email" className="block text-sm font-medium text-gray-700">
+              Email
+            </label>
+            <input
+              id="email"
+              type="email"
+              {...formik.getFieldProps('email')}
+              className={`mt-1 px-4 py-2 block w-full ${
+                formik.touched.email && formik.errors.email ? 'border-[1px] border-red-500' : 'border-none'
+              } rounded-md shadow-sm text-sm text-gray-800`}
+            />
+            {formik.touched.email && formik.errors.email && (
+              <p className="mt-2 text-xs text-red-600">{formik.errors.email}</p>
+            )}
+          </div>
+
+          {/* Password Input */}
+          <div className="mb-4 relative">
+            <label htmlFor="password" className="block text-sm font-medium text-gray-700">
+              Password
+            </label>
+            <input
+              id="password"
+              type={showPassword ? 'text' : 'password'}
+              {...formik.getFieldProps('password')}
+              className={`mt-1 px-4 py-2 block w-full ${
+                formik.touched.password && formik.errors.password ? 'border-[1px] border-red-500' : 'border-none'
+              } rounded-md shadow-sm text-sm text-gray-800`}
+            />
+            <div className={`absolute inset-y-10 right-0 pr-3 flex items-center text-sm leading-5 ${formik.touched.password && formik.errors.password ? 'inset-y-0' : 'inset-y-[42px]'}`}>
+              <button type="button" onClick={togglePasswordVisibility}>
+                {showPassword ? <EyeSlash size={20} /> : <Eye size={20} />}
+              </button>
+            </div>
+            {formik.touched.password && formik.errors.password && (
+              <p className="mt-2 text-xs text-red-600">{formik.errors.password}</p>
+            )}
+          </div>
+
+          {/* Submit Button */}
+          <div className="flex flex-col gap-2">
+            <button
+              type="submit"
+              className="text-white bg-stone-900 hover:bg-stone-950 font-medium rounded-lg text-sm px-5 py-2.5 text-center"
+            >
+              Login
             </button>
           </div>
-          <a href="#" className="text-xs flex justify-end mt-2 text-gray-800">
-            Forgot password?
-          </a>
-        </div>
-        <div className="flex flex-col gap-2">
-          <button
-            onClick={handleEmailLogin}
-            className="text-white bg-stone-900 hover:bg-stone-950 font-medium rounded-lg text-sm px-5 py-2.5 text-center"
-          >
-            Login
-          </button>
-        </div>
+        </form>
         <div>
           <p className="text-left text-gray-800 mt-4 text-xs">
-            Don&apos;t have an account? 
-            <Link href='register'><b> Register</b></Link>
+            Don&apos;t have an account?
+            <Link href="register">
+              <span className="font-inter font-bold text-black"> Register</span>
+            </Link>
           </p>
         </div>
-        {visibleInvalidFields && (
-          <label
-            htmlFor="invalidfields"
-            className="block text-sm font-medium text-gray-700"
-          >
-            {errorMessage}
-          </label>
+        {errorMessages !== "" && (
+          <div className="mt-4 p-2 bg-red-100 border-l-4 border-red-500">
+              <p className="text-red-600 text-sm">{errorMessages}</p>
+          </div>
         )}
       </div>
+      
     </div>
+    
   );
 };
 
 export default Login;
+
+
+
+
+
